better error handling

+++ src/monk/agent/string_operations.h
@@ -28,7 +28,7 @@ typedef struct {
 
 GArray* tokenize(const char* inputString, const char* delimiters);
 
-int streamTokenize(const char* inputChunk, int inputSize, const char* delimiters,
+int streamTokenize(const char* inputChunk, size_t inputSize, const char* delimiters,
         GArray** output, Token** remainder);
 
 #define tokenEquals(a, b) (((Token*) a)->hashedContent == ((Token*) b)->hashedContent)
